Clean up widget names and references
