+++ src/components/widgets/ColorWidget.js
@@ -1,5 +1,5 @@
 import React, {PropTypes} from "react";
-import BaseInput from "./EPBCBaseInput";
+import BaseInput from "./BaseInput";
 
 function ColorWidget(props) {
   return <BaseInput type="color" {...props}/>;
