+++ src/components/widgets/EmailWidget.js
@@ -1,5 +1,5 @@
 import React, {PropTypes} from "react";
-import BaseInput from "./EPBCBaseInput"
+import BaseInput from "./BaseInput"
 
 function EmailWidget(props) {
   return <BaseInput type="email" {...props}/>;
