+++ src/components/widgets/PasswordWidget.js
@@ -1,5 +1,5 @@
 import React, {PropTypes} from "react";
-import BaseInput from "./EPBCBaseInput";
+import BaseInput from "./BaseInput";
 
 function PasswordWidget(props) {
   return <BaseInput type="password" {...props}/>;
