+++ src/components/widgets/SelectWidget.js
@@ -1,6 +1,6 @@
-import React, {PropTypes} from "react";
+import React, { PropTypes } from "react";
 
-import {asNumber} from "../../utils";
+import { asNumber } from "../../utils";
 
 
 /**
@@ -9,11 +9,27 @@ import {asNumber} from "../../utils";
  */
 function processValue({type, items}, value) {
   if (type === "array" && items && ["number", "integer"].includes(items.type)) {
+    // If no selections have been made, return null...
+    if (value.length === 0) {
+      return null;
+    }
     return value.map(asNumber);
   } else if (type === "boolean") {
+    // If no selection has been made, return null...
+    if (value === "") {
+      return null;
+    }
     return value === "true";
-  } else if (type === "number") {
+  } else if (["number", "integer"].includes(type)) {
+    // If no selection has been made, return null...
+    if (value === 0) {
+      return null;
+    }
     return asNumber(value);
+  }
+  // If no selection has been made, return null...
+  if (value === "") {
+    return null;
   }
   return value;
 }
@@ -29,7 +45,8 @@ function SelectWidget({
   multiple,
   autofocus,
   ariaDescribedBy,
-  onChange
+  onChange,
+  onBlur
 }) {
   const {enumOptions} = options;
   return (
@@ -37,7 +54,7 @@ function SelectWidget({
       id={id}
       multiple={multiple}
       className="form-control"
-      value={value}
+      value={value == null ? "" : value}
       required={required}
       disabled={disabled}
       readOnly={readonly}
@@ -52,11 +69,21 @@ function SelectWidget({
           newValue = event.target.value;
         }
         onChange(processValue(schema, newValue));
-      }}>{
-      enumOptions.map(({value, label}, i) => {
-        return <option key={i} value={value}>{label}</option>;
-      })
-    }</select>
+      } }
+      onBlur={onBlur ? onBlur : (event) => {
+        let newValue;
+        if (multiple) {
+          newValue = [].filter.call(
+            event.target.options, o => o.selected).map(o => o.value);
+        } else {
+          newValue = event.target.value;
+        }
+        onChange(processValue(schema, newValue));
+      } }>{
+        enumOptions.map(({value, label}, i) => {
+          return <option key={i} value={value == null ? "" : value}>{label}</option>;
+        })
+      }</select>
   );
 }
 
@@ -77,6 +104,7 @@ if (process.env.NODE_ENV !== "production") {
     autofocus: PropTypes.bool,
     ariaDescribedBy: PropTypes.string,
     onChange: PropTypes.func,
+    onBlur: PropTypes.func,
   };
 }
