+++ src/components/widgets/TextWidget.js
@@ -1,6 +1,6 @@
 import React, {PropTypes} from "react";
 
-import BaseInput from "./EPBCBaseInput";
+import BaseInput from "./BaseInput";
 
 function TextWidget(props) {
   return <BaseInput {...props}/>;
