+++ src/components/widgets/URLWidget.js
@@ -1,5 +1,5 @@
 import React, {PropTypes} from "react";
-import BaseInput from "./EPBCBaseInput";
+import BaseInput from "./BaseInput";
 
 function URLWidget(props) {
   return <BaseInput type="url" {...props}/>;
