+++ src/components/widgets/index.js
@@ -4,18 +4,18 @@ import CheckboxWidget from "./CheckboxWidget";
 import CheckboxesWidget from "./CheckboxesWidget";
 import ConsentWidget from "./ConsentWidget";
 import ColorWidget from "./ColorWidget";
-import DateWidget from "./EPBCDateWidget";
+import DateWidget from "./DateWidget";
 import DateTimeWidget from "./DateTimeWidget";
 import DisplayWidget from "./DisplayWidget";
 import EmailWidget from "./EmailWidget";
 import FileWidget from "./FileWidget";
 import HiddenWidget from "./HiddenWidget";
-import MonthYearWidget from "./EPBCMonthYearWidget";
+import MonthYearWidget from "./MonthYearWidget";
 import PasswordWidget from "./PasswordWidget";
 import RadioWidget from "./RadioWidget";
 import RangeWidget from "./RangeWidget";
-import SelectWidget from "./EPBCSelectWidget";
-import TextareaWidget from "./EPBCTextareaWidget";
+import SelectWidget from "./SelectWidget";
+import TextareaWidget from "./TextareaWidget";
 import TextWidget from "./TextWidget";
 import UpDownWidget from "./UpDownWidget";
 import URLWidget from "./URLWidget";
